+++ src/constants.js
@@ -19,16 +19,17 @@ const WEB_VITALS_ACCESSOR_KEY = "__cy_web_vitals__";
 const WEB_VITALS_SNIPPET = `<script type="module">
 import {getCLS,getFID,getLCP,getFCP,getTTFB} from "https://unpkg.com/web-vitals?module";
 window.${WEB_VITALS_ACCESSOR_KEY}=window.${WEB_VITALS_ACCESSOR_KEY}||{};
-getCLS(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.cls=m;console.log(m);});
-getFID(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.fid=m;console.log(m);});
-getLCP(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.lcp=m;console.log(m);});
-getFCP(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.fcp=m;console.log(m);});
-getTTFB(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.ttfb=m;console.log(m);});
+const event = new Event("${WEB_VITALS_ACCESSOR_KEY}");
+getCLS(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.cls=m;console.log(m);window.dispatchEvent(event);});
+getFID(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.fid=m;console.log(m);window.dispatchEvent(event);});
+getLCP(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.lcp=m;console.log(m);window.dispatchEvent(event);});
+getFCP(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.fcp=m;console.log(m);window.dispatchEvent(event);});
+getTTFB(m=>{window.${WEB_VITALS_ACCESSOR_KEY}.ttfb=m;console.log(m);window.dispatchEvent(event);});
 </script>`;
 
 const LOG_SLUG = "cy.vitals()";
 
-const DELAY_MS_FOR_CLS_REPORTING = 100;
+const ALL_WEB_VITALS_REPORTED_TIMEOUT_MS = 4000;
 
 module.exports = {
   DEFAULT_THRESHOLDS,
@@ -37,5 +38,5 @@ module.exports = {
   WEB_VITALS_ACCESSOR_KEY,
   WEB_VITALS_SNIPPET,
   LOG_SLUG,
-  DELAY_MS_FOR_CLS_REPORTING,
+  ALL_WEB_VITALS_REPORTED_TIMEOUT_MS,
 };